requirements.txt
fastapi
uvicorn
celery
redis
yt-dlp
python-dotenv
pydantic
pillow
sqlalchemy
psycopg2-binary